Chore: Bump Spark SDK version
