+++ examples/with-spark/package.json
@@ -13,7 +13,7 @@
   },
   "dependencies": {
     "@buildonspark/issuer-sdk": "^0.1.23",
-    "@buildonspark/spark-sdk": "0.8.4",
+    "@buildonspark/spark-sdk": "0.8.6",
     "@noble/curves": "^1.9.7",
     "@noble/hashes": "^1.4.0",
     "@scure/bip39": "^1.3.0",
