+++ packages/spark/package.json
@@ -47,7 +47,7 @@
     "typecheck": "tsc --noEmit"
   },
   "dependencies": {
-    "@buildonspark/spark-sdk": "^0.8.4",
+    "@buildonspark/spark-sdk": "^0.8.6",
     "@noble/curves": "^1.9.7",
     "@noble/hashes": "^1.4.0",
     "@scure/bip39": "^1.3.0",
