+++ packages/spark/src/services/index.ts
@@ -1,1 +1,6 @@
+export {
+  TurnkeyCoopExitService,
+  type GetConnectorRefundSignaturesParams,
+  type GetConnectorRefundSignaturesReturn,
+} from "./coop-exit";
 export { TurnkeyTransferService } from "./transfer";
